=act split ReflectiveDynamicAccess implementation in separate file.
